Implement percpu blocks/variables, use for syscalls
Each CPU gets its own cpu_t object which contains metadata about
the currently executing process and references to per-CPU variable
blocks. The first CPU is statically allocated, subsequent ones will
be dynamically allocated.

Per-CPU variables are implemented by having a 'reference' frame in the
ELF statically created with all per-cpu variables. The offset of each
variable into the reference frame is used as an offset into each dynamic
per CPU region to access per-cpu variables.

On i386/i686, the per-CPU block is stored in the GS register. For i386,
this is loaded via TSS on syscall entry; on i686, the swapgs instruction
is used to load the block from an MSR.