x86_64: Ensure that the pages reserved for Multiboot metadata are mapped
The Multiboot spec doesn't have anything to say about where it will
place the Multiboot metadata; in some versions and some kernel layouts,
this data is placed at a higher address than the page in which the kernel end
resides.

Fix the standup page mappings to include the Multiboot info struct, as
well as any ELF metadata that is required (i.e. string symbols).